requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
packaging==24.1
pillow==10.4.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
requests==2.32.3
setuptools==75.1.0
tkhtmlview==0.3.1
urllib3==2.2.3
